+++ src/string/string_1008.rs
@@ -282,7 +282,7 @@ fn unk_str_op_1008_d1c6(param_1: u32,param_2: u32)
       uVar17 = 0x1018;
       pass1_1018_4808(CONCAT22(puVar14,uVar9),0x0,
                       uStack14 & 0xffff | ZEXT24(puVar13) << 0x10,0x0);
-      uVar10 = (u8)uVar15;
+      uVar10 = uVar15;
     }
     uVar2 = (iVar16 + 0x12);
     ppcVar3 = ((iVar16 + 0x12) + 0x4);
@@ -412,7 +412,7 @@ fn unk_str_op_1008_d4f6(param_1: u32,param_2: u32)
       struct_1018_47c8(CONCAT22(puVar18,uVar10),0x0,
                        CONCAT13((puVar14 >> 0x8),
                                 CONCAT12(puVar14,uVar12)),0x0,0x0);
-      uVar13 = (u8)uVar19;
+      uVar13 = uVar19;
     }
     uVar3 = (iVar21 + 0xe);
     ppcVar4 = ((iVar21 + 0xe) + 0x4);
